[3.1] Create HttpServerTracerBase decorator for manual instrumentation
