+++ settings.gradle
@@ -151,6 +151,7 @@ include ':instrumentation:vertx-3.0'
 include ':instrumentation:vertx-reactive-3.5'
 
 include ':instrumentation-core:aws-sdk:aws-sdk-2.2'
+
 include ':instrumentation-core:spring'
 
 // exporter adapters
